Update Travis API URL from `.org` to `.com`
This is in anticipation of the mandatory [migration from `travis-ci.org` to `travis-ci.com`](https://docs.travis-ci.com/user/migrate/open-source-repository-migration#migrating-a-repository) that was announced in 2018, but is being enforced soon. AMP is going to do this migration later today. I'll merge the PR after the migration is in place. This PR is not in the critical path, so it can be reviewed and merged when convenient.

Companion PR to https://github.com/ampproject/amphtml/pull/30846